Added the explanation for database terms.
Added an image for to show the common parts of prepared and parametrized statement. Also the `docs.html` which the idea is to give links to the configuration and the plain documentation for the frameworks, so you may see what a specific setting does or when you are looking for how can you configure it, so that you may not search it all over again.